requirements.txt
scipy==1.4.1
opencv_contrib_python==4.4.0.46
streamlit==0.72.0
streamlit_cropper==0.1.3
matplotlib==3.2.2
numpy==1.18.4
scikit_image==0.16.2
Pillow==8.0.1
skimage==0.0
